Separate out db functionalities from the rest of indra.
