--- indra/db/__init__.py
@@ -1,2 +1,0 @@
-from .database_manager import texttypes, formats, sql_expressions
-from .util import get_primary_db
