+++ indra_db/managers/content_manager.py
@@ -109,9 +109,9 @@ from indra.util import UnicodeXMLTreeBuilder as UTB
 from indra.literature.pmc_client import id_lookup
 from indra.literature import pubmed_client
 
-from indra.db.util import get_primary_db, get_db, get_test_db
-from indra.db.database_manager import texttypes, formats
-from indra.db.database_manager import sql_expressions as sql_exp
+from indra_db.util import get_primary_db, get_db, get_test_db
+from indra_db.managers.database_manager import texttypes, formats
+from indra_db.managers.database_manager import sql_expressions as sql_exp
 
 
 try:
