+++ indra_db/resources/default_db_config.ini
